requirements.txt
asgiref==3.9.1
Django==5.2.6
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
anyio==4.10.0
asgiref==3.9.1
blinker==1.9.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.8.3
charset-normalizer==3.4.3
cleo==2.1.0
click==8.2.1
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.1
packaging==25.0
pbs-installer==2025.8.8
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.4
poetry-core==2.1.3
pyproject_hooks==1.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.8.6.13
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.32.0
Werkzeug==3.1.3
zstandard==0.23.0
